pyproject.toml
[project]
name = "ethical-responsible-gen-ai"
version = "0.1.0"
description = "Example of ethical and responsible use of LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.5.0",
    "google-genai>=1.13.0",
    "llm-guard>=0.3.15",
    "python-dotenv>=1.1.0",
]
